refactor: BeanScanner 기능 분리
