+++ src/main/java/core/mvc/tobe/AnnotationHandlerMapping.java
@@ -1,30 +1,60 @@
 package core.mvc.tobe;
 
 import com.google.common.collect.Maps;
+import core.annotation.web.RequestMapping;
 import core.annotation.web.RequestMethod;
+import core.di.factory.BeanFactory;
 import core.mvc.HandlerMapping;
+import core.mvc.tobe.support.*;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.core.LocalVariableTableParameterNameDiscoverer;
+import org.springframework.core.ParameterNameDiscoverer;
 
 import javax.servlet.http.HttpServletRequest;
+import java.lang.reflect.Method;
+import java.util.Arrays;
+import java.util.List;
 import java.util.Map;
+
+import static java.util.Arrays.asList;
 
 public class AnnotationHandlerMapping implements HandlerMapping {
     private static final Logger logger = LoggerFactory.getLogger(AnnotationHandlerMapping.class);
+    private static final ParameterNameDiscoverer nameDiscoverer = new LocalVariableTableParameterNameDiscoverer();
+    private static final List<ArgumentResolver> argumentResolvers = asList(
+        new HttpRequestArgumentResolver(),
+        new HttpResponseArgumentResolver(),
+        new RequestParamArgumentResolver(),
+        new PathVariableArgumentResolver(),
+        new ModelArgumentResolver()
+    );
 
-    private Object[] basePackage;
-    private BeanScanner beanScanner;
-
-    private Map<HandlerKey, HandlerExecution> handlerExecutions = Maps.newHashMap();
+    private final Object[] basePackage;
+    private final Map<HandlerKey, HandlerExecution> handlerExecutions = Maps.newHashMap();
 
     public AnnotationHandlerMapping(Object... basePackage) {
         this.basePackage = basePackage;
-        beanScanner = new BeanScanner();
     }
 
     public void initialize() {
         logger.info("## Initialized Annotation Handler Mapping");
-        handlerExecutions.putAll(beanScanner.scan(basePackage));
+        BeanScanner beanScanner = new BeanScanner();
+        BeanFactory beanFactory = new BeanFactory(beanScanner.scan(basePackage));
+        Map<Class<?>, Object> controllers = beanFactory.getControllers();
+        controllers.forEach((clazz, handler) -> addHandlerExecution(handlerExecutions, handler, clazz.getMethods()));
+    }
+
+    private void addHandlerExecution(Map<HandlerKey, HandlerExecution> handlers, final Object target, Method[] methods) {
+        Arrays.stream(methods)
+            .filter(method -> method.isAnnotationPresent(RequestMapping.class))
+            .forEach(method -> {
+                RequestMapping requestMapping = method.getAnnotation(RequestMapping.class);
+                HandlerKey handlerKey = new HandlerKey(requestMapping.value(), requestMapping.method());
+                HandlerExecution handlerExecution = new HandlerExecution(nameDiscoverer, argumentResolvers, target, method);
+                handlers.put(handlerKey, handlerExecution);
+                logger.info("Add - method: {}, path: {}, HandlerExecution: {}", requestMapping.method(), requestMapping.value(), method.getName());
+            });
     }
 
     public Object getHandler(HttpServletRequest request) {
